+++ src/app/[locale]/post-security/clue-item.tsx
@@ -10,6 +10,7 @@ import {
 } from "@/components/ui/accordion";
 import { Card, CardContent } from "@/components/ui/card";
 import { AnswerType, Clue, ClueType } from "@/types/clue";
+import { useTranslations } from "next-intl";
 import Image from "next/image";
 
 export interface ClueItemProperties {
@@ -17,16 +18,18 @@ export interface ClueItemProperties {
 }
 
 export function ClueItem({ clue }: ClueItemProperties) {
+  const t = useTranslations("clues");
+
   return (
     <Card className="p-0" aria-labelledby={`clue-${clue.id}`} role="group">
       <CardContent className="px-4 py-3 space-y-2">
         <h3 id={`clue-${clue.id}`} className="pb-2 font-bold !text-base">
-          {clue.clue}
+          {t(`${clue.id}.clue`)}
         </h3>
         {clue.clueType === ClueType.IMAGE && (
           <div className="relative mx-auto max-w-[400px] aspect-video">
             <Image
-              alt={clue.alternateText}
+              alt={t(`${clue.id}.alternateText`)}
               src={`/clue-images/${clue.id}.jpg`}
               fill
               sizes="(max-width: 420px) 100vw, 400px"
@@ -34,21 +37,21 @@ export function ClueItem({ clue }: ClueItemProperties) {
             />
           </div>
         )}
-        {clue.hint && (
+        {t.has(`${clue.id}.hint`) && (
           <Markdown className="text-muted-foreground mb-2">
-            {clue.hint}
+            {t(`${clue.id}.hint`)}
           </Markdown>
         )}
         {clue.answerType === AnswerType.TEXT && <TextAnswer id={clue.id} />}
         {clue.answerType === AnswerType.IMAGE && <ImageAnswer id={clue.id} />}
-        {(clue.answer || clue.answerDetails) && (
+        {(t.has(`${clue.id}.answer`) || t.has(`${clue.id}.answerDetails`)) && (
           <Accordion type="single" collapsible className="w-full">
             <AccordionItem value="answer">
               <AccordionTrigger className="pb-0">Answer</AccordionTrigger>
               <AccordionContent className="pb-0 mt-2">
-                {clue.answer && <p>{clue.answer}</p>}
-                {clue.answerDetails && (
-                  <Markdown>{clue.answerDetails}</Markdown>
+                {t.has(`${clue.id}.answer`) && <p>{t(`${clue.id}.answer`)}</p>}
+                {t.has(`${clue.id}.answerDetails`) && (
+                  <Markdown>{t(`${clue.id}.answerDetails`)}</Markdown>
                 )}
               </AccordionContent>
             </AccordionItem>
