+++ src/data/post-security-clues.ts
@@ -1,0 +1,280 @@
+import { AirportArea, AnswerType, Clue, ClueType } from "@/types/clue";
+
+export const postSecurityClues: ReadonlyArray<Clue> = [
+  {
+    id: "XU2Q6uk5Sj", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_B,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "Fq6cpEvGKt", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_B,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "BL5xfCD1Pl", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "KFNxBoixx9", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_B,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "JTaK-wv-cm", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "GpyOC1fxZv", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "WOflQcqqRa", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "wIuAhUYSUJ", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "Mr8NTBLVuw", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "Y7qiTGjFDA", // cspell: disable-line
+    airportArea: AirportArea.SOUTH_SATELLITE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "CEDiAt5_TD", // cspell: disable-line
+    airportArea: AirportArea.SOUTH_SATELLITE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "mdihls1FXP", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_C,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "qsWmZtJQx9", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "vVzEujhDHt", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_C,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "fId26KufqS", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "vVonLvfgvX", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "Z8uD5RIdFk", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "U2Ml_N2Ao1", // cspell: disable-line
+    airportArea: AirportArea.CENTRAL_TERMINAL,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "uaPC_HvMqd", // cspell: disable-line
+    airportArea: AirportArea.CENTRAL_TERMINAL,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "1XWgzKECE6", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "xH4xrNHYJe", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "pB53_yVOPQ", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "96FePKs7zD", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "ShfXCzAYKC", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_C,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "4B8BuOoy-A", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "cvI9vHSRuC", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "PWYzT5xEM_", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "FwkTyDghkR", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT, // hint: "",
+  },
+  {
+    id: "VGDb3ekZZs", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_A,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "yTNP29jtdB", // cspell: disable-line
+    airportArea: AirportArea.NORTH_SATELLITE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "-buCfuZ_Js", // cspell: disable-line
+    airportArea: AirportArea.CENTRAL_TERMINAL,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "JjwwVsnaNJ", // cspell: disable-line
+    airportArea: AirportArea.SOUTH_SATELLITE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "ewJw9uhuxK", // cspell: disable-line
+    airportArea: AirportArea.CENTRAL_TERMINAL,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "ndC0lIJES6", // cspell: disable-line
+    airportArea: AirportArea.NORTH_SATELLITE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "v86d2mm7UT", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "6IjATuH8h-", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_B,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "nGRX7LLH2E", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_C,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "glwn2opFlc", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_C,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "uzxCpAxD0O", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "Ajg7X3_CWe", // cspell: disable-line
+    airportArea: AirportArea.CENTRAL_TERMINAL,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "tYEDPlu30G", // cspell: disable-line
+    airportArea: AirportArea.SOUTH_SATELLITE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "sZLbWTNUXK", // cspell: disable-line
+    airportArea: AirportArea.SOUTH_SATELLITE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "-GLJeZ-L-K", // cspell: disable-line
+    airportArea: AirportArea.CONCOURSE_D,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "xNr6O90eQa", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.TEXT,
+    clueType: ClueType.TEXT,
+  },
+  {
+    id: "MSjgrbHme1", // cspell: disable-line
+    clueType: ClueType.IMAGE,
+    airportArea: AirportArea.NORTH_SATELLITE,
+    answerType: AnswerType.IMAGE,
+  },
+  {
+    id: "a8vW4xzkkW", // cspell: disable-line
+    airportArea: AirportArea.AIRPORT_WIDE,
+    answerType: AnswerType.IMAGE,
+    clueType: ClueType.TEXT,
+  },
+];
